+++ src/extension.ts
@@ -1,17 +1,23 @@
-import { commands, ExtensionContext, languages, workspace } from "vscode";
+import * as vscode from "vscode";
 import { generateAssets } from "./commands/generateAssets/command";
 import { addProjectToSolution } from "./commands/solution/add/command";
 import { handleNewFile } from "./features/newFile";
 
-const _ = "csharp-utilities.";
+export const ns = "csharp-utilities";
 
-export async function activate(context: ExtensionContext) {
+export async function activate(context: vscode.ExtensionContext) {
     // Commands
-    context.subscriptions.push(commands.registerCommand(_ + "generate-assets", generateAssets));
-    context.subscriptions.push(commands.registerCommand(_ + "solution.add", addProjectToSolution));
+    context.subscriptions.push(vscode.commands.registerCommand(ns + ".generate-assets", generateAssets));
+    context.subscriptions.push(vscode.commands.registerCommand(ns + ".solution.add", addProjectToSolution));
+
+    let autoGenerateAssets = vscode.workspace.getConfiguration(ns).get<Boolean>("autoGenerateAssets")
+
+    if (autoGenerateAssets) {
+        generateAssets();
+    }
 
     // Features
-    context.subscriptions.push(workspace.onDidCreateFiles(handleNewFile));
+    context.subscriptions.push(vscode.workspace.onDidCreateFiles(handleNewFile));
 }
 
 export function deactivate() { }